+++ src/containers/Drawers/ItemDetailDrawer/ItemDetailDrawerProvider.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { DrawerHeaderContent, DrawerLoading } from 'components';
 import { useItem } from 'hooks/query';
+import { inactiveStatus } from './utlis';
 
 const ItemDetailDrawerContext = React.createContext();
 
@@ -27,7 +28,10 @@ function ItemDetailDrawerProvider({ itemId, ...props }) {
 
   return (
     <DrawerLoading loading={isItemLoading}>
-      <DrawerHeaderContent name="item-detail-drawer" title={item?.name} />
+      <DrawerHeaderContent
+        name="item-detail-drawer"
+        title={inactiveStatus(item)}
+      />
       <ItemDetailDrawerContext.Provider value={provider} {...props} />
     </DrawerLoading>
   );