+++ src/hooks/query/index.js
@@ -33,5 +33,6 @@ export * from './roles';
 export * from './creditNote';
 export * from './vendorCredit';
 export * from './transactionsLocking';
-export * from './warehouses'
+export * from './warehouses';
 export * from './branches';
+export * from './warehousesTransfers';